+++ connector_ebisumart/models/__init__.py
@@ -1,5 +1,5 @@
 # from . import sale_order
-# from . import automatic_workflow_job
+from . import automatic_workflow_job
 from . import stock_rule
 from . import res_partner
 from . import account_journal
