+++ connector_ebisumart/models/automatic_workflow_job.py
@@ -7,32 +7,66 @@ from odoo import models
 class AutomaticWorkflowJob(models.Model):
     _inherit = 'automatic.workflow.job'
 
-    def _do_validate_sale_order(self, sale):
-        super()._do_validate_sale_order(sale)
-        # Search purchase order
-        po = self.env['purchase.order'].search([('origin', '=', sale.name)])
-        if not po:
-            return
-        # map data_planned
-        for line in po.order_line:
-            line.date_planned = po.date_order
-        # Confirm the purchase order.
-        po.button_confirm()
-        # Validate the associated receipt (stock picking).
-        for picking in po.picking_ids:
-            wiz = self.env['stock.immediate.transfer'].create(
-                {'pick_ids': [(4, picking.id)]}
-            )
-            wiz.process()
+    def _do_validate_picking(self, picking):
+        super()._do_validate_picking(picking)
+        if picking.sale_id.cancel_in_ebisumart:
+            self.create_return_picking(picking.sale_id)
+            purchase_order = self.env['purchase.order'].search([('origin', '=', picking.sale_id.name)])
+            if not purchase_order:
+                return
+            self.create_return_picking(purchase_order)
 
-        po_invoice = {
-            'partner_id': po.partner_id.id,
-            'account_id': po.partner_id.property_account_payable_id.id,
-            'state': 'draft',
-            'type': 'in_invoice',
-            'purchase_id': po.id,
-        }
+    def _do_validate_invoice(self, invoice):
+        super()._do_validate_invoice(invoice)
+        sale_order = invoice.mapped('invoice_line_ids.sale_line_ids.order_id')
+        if sale_order.cancel_in_ebisumart:
+            self.create_credit_note(sale_order, invoice_type="out_invoice")
+            purchase_order = self.env['purchase.order'].search([('origin', '=', sale_order.name)])
+            if not purchase_order:
+                return
+            self.create_credit_note(purchase_order, invoice_type="in_invoice")
 
-        inv = self.env['account.invoice'].create(po_invoice)
-        inv.purchase_order_change()
-        inv.action_invoice_open()
+    def create_return_picking(self, order):
+        for picking in order.picking_ids.filtered(lambda r: r.state == 'done'):
+            # Create the reverse transfer
+            stock_return_picking = self.env['stock.return.picking']
+            return_wizard = stock_return_picking.with_context(
+                active_ids=picking.ids, active_id=picking.ids[0]
+            ).create({})
+            return_result = return_wizard.create_returns()
+
+            # Usually, the return_result contains information
+            # about the newly created return picking(s)
+            if return_result and 'res_id' in return_result:
+                new_return_picking = self.env['stock.picking'].browse(
+                    return_result['res_id']
+                )
+
+                # Validate (confirm) the return picking
+                if new_return_picking.state != 'done':
+                    wiz = self.env['stock.immediate.transfer'].create({
+                        'pick_ids': [(4, new_return_picking.id)]
+                    })
+                    wiz.process()
+
+    def create_credit_note(self, order, invoice_type):
+        for invoice in order.invoice_ids.filtered(
+            lambda r: r.state not in ['cancel', 'draft'] and r.type == invoice_type
+        ):
+            # Create the refund (credit note)
+            account_invoice_refund = self.env['account.invoice.refund']
+            refund_wizard = account_invoice_refund.create({
+                'description': 'Credit Note',
+                'filter_refund': 'refund',  # refund the entire invoice
+            })
+            refund_result = refund_wizard.with_context(
+                active_ids=invoice.ids
+            ).invoice_refund()
+            if refund_result and refund_result.get('domain'):
+                # Search for the newly created credit note
+                credit_notes = self.env['account.invoice'].search(
+                    refund_result.get('domain')
+                )
+                for credit_note in credit_notes:
+                    if credit_note.state == 'draft':
+                        credit_note.action_invoice_open()
